+++ HISTORY.md
@@ -2,6 +2,10 @@
 
 ## in development:
 
+### 2020-08-04
+* new dev version
+* blocks: include background color when exporting (semi-) transparent script pics 
+
 ## 6.1.2:
 * fixed variable scope for ASK/TELL